refactor(tls): extract fingerprint tests to sibling file
